requirements.txt
beautifulsoup4
requests
pydantic
click
jinja2
staticjinja
sqlite-utils
python-dateutil
python-dotenv